requirements.txt
matplotlib==3.5.3
numpy==1.23.1
PyYAML==6.0
requests==2.22.0
scipy==1.9.1
torch==1.7.0
torchmetrics==0.9.3
varname==0.10.0
asttokens
